moved base from server to common
