+++ vantage6-server/vantage6/server/model/__init__.py
@@ -2,7 +2,7 @@ from vantage6.server.model.authenticatable import Authenticatable
 from vantage6.server.model.user import User
 from vantage6.server.model.organization import Organization
 from vantage6.server.model.collaboration import Collaboration
-from vantage6.server.model.base import Base
+from vantage6.common.base import Base
 from vantage6.server.model.member import Member, StudyMember
 from vantage6.server.model.node_config import NodeConfig
 from vantage6.server.model.node import Node
