+++ vantage6-server/vantage6/server/model/algorithm_store.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from sqlalchemy import Column, String, Integer, ForeignKey
 from sqlalchemy.orm import relationship
 
-from .base import Base, DatabaseSessionManager
+from vantage6.common.base import Base, DatabaseSessionManager
 
 
 class AlgorithmStore(Base):
