+++ vantage6-server/vantage6/server/model/task_database.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, Integer, ForeignKey, String
 from sqlalchemy.orm import relationship
 
-from .base import Base
+from vantage6.common.base import Base
 
 
 class TaskDatabase(Base):
